feat: fix import and refactor structure
