+++ src/agents/Critic.ts
@@ -1,7 +1,7 @@
 import { execa } from 'execa';
 import { to } from 'await-to-js';
 import { v4 as uuid } from 'uuid';
-import { KnowledgeGraphManager, DagNode } from '../KnowledgeGraph.ts';
+import { KnowledgeGraphManager, DagNode } from '../engine/KnowledgeGraph.ts';
 import { RevisionCounter } from '../actor-critic/RevisionCounter.ts';
 import path from 'node:path';
 import { fileURLToPath } from 'node:url';
